Move HardContextModuleFactory creation to a plugin
